backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Cors==5.0.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.0
MarkupSafe==3.0.2
multidict==6.1.0
openai==0.28.0
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
pytest
pytest-flask
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.2
yarl==1.17.1
